Allow specifying xPad and yPad individually.
For the group methods, there is two types of padding: between subviews and between the group and its sibling or superview. In all the grouping methods (except groupInCenter), we want to allow specifying the xPad and yPad values individually, so that the user can control both the padding between subviews and the padding between the group and its superview.